Revert "Some Optimizations"
This reverts commit 7a15fa90031ee63c762910972de40717b4acd407.

+++ bb_common/src/math/fpos.rs
@@ -127,7 +127,7 @@ impl FPos {
 
   /// Returns the distance to the other position.
   pub fn dist(&self, other: FPos) -> f64 {
-    ((self.x - other.x).powi(2) + (self.y - other.y).powi(2) + (self.z - other.z).powi(2))
+    (((self.x - other.x).powi(2) + (self.y - other.y).powi(2) + (self.z - other.z).powi(2)) as f64)
       .sqrt()
   }
   /// Returns the squared distance to the other position. Since block positions
